+++ Assets/Scripts/Controller/UserDataController.cs
@@ -5,9 +5,9 @@ namespace MyGameNamespace
 {
     public class UserDataController : MonoBehaviour, IUserData
     {
-        [field: SerializeField] public int Coin { get; set; }
-        [field: SerializeField] public int Diamond { get; set; }
-        [field: SerializeField] public UnityEvent<(int Before, int after)> OnCoinChange { get; set; }
-        [field: SerializeField] public UnityEvent<(int Before, int after)> OnDiamondChange { get; set; }
+        [field: SerializeField] public long Coin { get; set; }
+        [field: SerializeField] public long Diamond { get; set; }
+        [field: SerializeField] public UnityEvent<(long Before, long after)> OnCoinChange { get; set; }
+        [field: SerializeField] public UnityEvent<(long Before, long after)> OnDiamondChange { get; set; }
     }
 }
